Develop -> feature/Question (#35)
* :bug: merge 위한 코드 수정

* :bug: merge 위한 코드 수정 (1)

* Feature/auth2 - 예외 처리, 코드 수정 (#20)

* test

* :yellow_heart: refresh token 만료시 재생성

* :bug: userId -> userEntity 변경, 중복 회원가입 작성중

* :bug: userEntity 도입 추가 수정

* :bug: 회원가입 중복 체크, 서비스 Impl 분리

* ✨ Implemented Answer API (Connect to Auth) (#21)

* Remove Interface

* Added FORBIDEN 403 Exception

* Implemented Exceptions

* Modify some exception

* Added select & unselect answer

* Change throwing exception from selected to closed while update/delete

* Implemented Getting Answers

* Try to apply authentication, wait for fix

* Added TODO (About useranswerinteraction)

* Fixed duplicated attempt to save on db when create answer

* 🔧 Changed jwt expiration time to 5 min, ddl auto to create on prod profile

* 🔧 Changed jwt expiration time to 5 min, ddl auto to create on prod profile

* 🔀 Merge Develop (Temporary profile change for db creation) (#22)

* 🔧 Revert to prod profile do nothing about database

* ✨ Implemented UserAnswerInteraction API (#24)

* 🚑️ Fixed Question, Answer Entity to Save Order of Photos (#25)

* Change Photo OneToMany relation using MutableSet, and added photosOrder field in PhotoEntity

* Fixed Services according to changes in db

* Fix linting

* 🏷️ Add id of answer and user in AnswerResponse (#26)

* Added id and user id in answerresponse

* Make delete answer to delete useranswerinteractions

* Feature/auth2.1 Add checkDuplicatedId, logout, getKakaoToken, kakaoLogin (#28)

* test

* :sparkles: 카카오 서비스 일부 생성, request 디렉토리 이동

* :sparkles: 로그아웃 생성

* :sparkles: 회원가입 내 중복체크 생성

* :yellow_heart: 카카오 로그인 구체화

* :sparkles: 프론트랑 연결 없이, 백에서 자체 카카오 토큰 생성

* :yellow_heart: 카카오 회원가입->회원정보 db 저장, 토큰 생성까지 완료

* :yellow_heart: 컨트롤러 불필요한거 삭제

* ktlint 수정

* ✨ Configured S3 Server, Test API for Image Upload/Delete, Hide JWT Configs (#29)

* Added application-secret.yaml as ignoring file

* Configured profile and libraries for s3 bucket

* Config for S3

* Created S3Service and test api for upload photo

* Fixed Typo

* Added S3 service for upload & remove file

* Added testing api for image upload & remove

* Move jwt secrets to secret profile

* 🚑️ Fix yaml config file typo

* Feature/auth3 getMyQuestions, queryDsl, regerateToken (#32)

* :yellow_heart: 드디어 Auth3 시작! ㅎㅎ

* :yellow_heart: 메서드 정리, getKakaoUserInfo를 private 메서드로 바꿈

* :yellow_heart: 메서드 2차 정리

* :yellow_heart: 메서드 3차 정리

* application-secret 실험중

* :yellow_heart: queryDsl 도입중

* queryDsl 작성중

* :yellow_heart: querydsl 작성중... Login response에 username 추가

* :sparkles: getMyQuestions 생성

* :sparkles: getMyAnswers 생성

* getMyAnswers 생성 수정

* ktlint 수정

* Delete application-secret.yaml

* gitignore 시도중

* gitignore 시도중

* gitignore 시도중2

* gitignore 시도중3

* :sparkles: queryDslRepository 삭제 및 이동, getMyAllProfile 생성

* getMyAllProfiles에서 answerNumber 추가

* :sparkles: authInteceptor 수정, regenerateToken 생성

+++ src/main/kotlin/com/wafflestudio/team2/jisik2n/core/answer/service/AnswerService.kt
@@ -1,18 +1,26 @@
 package com.wafflestudio.team2.jisik2n.core.answer.service
 
+import com.wafflestudio.team2.jisik2n.common.Jisik2n400
+import com.wafflestudio.team2.jisik2n.common.Jisik2n403
+import com.wafflestudio.team2.jisik2n.common.Jisik2n404
 import com.wafflestudio.team2.jisik2n.core.answer.database.AnswerEntity
 import com.wafflestudio.team2.jisik2n.core.answer.database.AnswerRepository
 import com.wafflestudio.team2.jisik2n.core.answer.dto.AnswerRequest
+import com.wafflestudio.team2.jisik2n.core.answer.dto.AnswerResponse
 import com.wafflestudio.team2.jisik2n.core.photo.database.PhotoEntity
 import com.wafflestudio.team2.jisik2n.core.photo.database.PhotoRepository
 import com.wafflestudio.team2.jisik2n.core.question.database.QuestionRepository
 import com.wafflestudio.team2.jisik2n.core.user.database.UserEntity
-import com.wafflestudio.team2.jisik2n.core.user.database.UserRepository
+import com.wafflestudio.team2.jisik2n.core.userAnswerInteraction.database.UserAnswerInteractionRepository
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
+import java.time.LocalDateTime
+import java.time.ZoneId
 import javax.transaction.Transactional
 
 interface AnswerService {
+    fun getAnswersOfQuestion(questionId: Long): List<AnswerResponse>
+
     fun createAnswer(
         loginUser: UserEntity,
         questionId: Long,
@@ -25,6 +33,8 @@ interface AnswerService {
         answerRequest: AnswerRequest
     )
 
+    fun toggleSelectAnswer(loginUser: UserEntity, answerId: Long, toSelect: Boolean)
+
     fun removeAnswer(loginUser: UserEntity, answerId: Long)
 }
 
@@ -33,8 +43,19 @@ class AnswerServiceImpl(
     private val answerRepository: AnswerRepository,
     private val questionRepository: QuestionRepository,
     private val photoRepository: PhotoRepository,
-    private val userRepository: UserRepository,
+    private val userAnswerInteractionRepository: UserAnswerInteractionRepository,
 ) : AnswerService {
+    override fun getAnswersOfQuestion(questionId: Long): List<AnswerResponse> {
+        // Get target question
+        val question = questionRepository.findByIdOrNull(questionId)
+            ?: throw Jisik2n404("${questionId}에 해당하는 질문이 없습니다.")
+
+        // TODO: Improve query
+        val answers = question.answers.sortedWith(compareBy({ !it.selected }, { it.createdAt }))
+
+        return answers.map { it.toResponse(answerRepository) }
+    }
+
     @Transactional
     override fun createAnswer(
         loginUser: UserEntity,
@@ -43,7 +64,11 @@ class AnswerServiceImpl(
     ) {
         // Get target question
         val question = questionRepository.findByIdOrNull(questionId)
-            ?: TODO("Throw 404 Exception")
+            ?: throw Jisik2n404("${questionId}에 해당하는 질문이 없습니다.")
+
+        if (loginUser.id == question.user.id) {
+            throw Jisik2n403("자신의 질문에는 답할 수 없습니다.")
+        }
 
         // Add new answer
         var newAnswer = answerRequest.let {
@@ -55,24 +80,12 @@ class AnswerServiceImpl(
         }
 
         // Add photos to newAnswer
-        answerRequest.photos.map { path: String ->
-            PhotoEntity(path, answer = newAnswer)
+        answerRequest.photos.mapIndexed { idx: Int, path: String ->
+            PhotoEntity(path, idx, answer = newAnswer)
         }.also {
             newAnswer.photos.addAll(it)
         }
         newAnswer = answerRepository.save(newAnswer)
-
-        // Add newAnswer to question
-        question.let {
-            it.answers.add(newAnswer)
-            questionRepository.save(it)
-        }
-
-        // Add newAnswer to user answers
-        loginUser.let {
-            it.answers.add(newAnswer)
-            userRepository.save(it)
-        }
     }
 
     @Transactional
@@ -82,10 +95,13 @@ class AnswerServiceImpl(
         answerRequest: AnswerRequest
     ) {
         val answer = answerRepository.findByIdOrNull(answerId)
-            ?: TODO("Throw 404 Not Found Exception")
+            ?: throw Jisik2n404("${answerId}에 해당하는 답변이 없습니다")
 
-        if (answer.user.id != loginUser.id || answer.selected) {
-            TODO("Throw 403 Forbidden Exception")
+        if (answer.user.id != loginUser.id) {
+            throw Jisik2n403("자신의 게시물만 수정할 수 있습니다.")
+        }
+        if (answer.question.close) {
+            throw Jisik2n403("마감된 질문은 수정될 수 없습니다.")
         }
 
         // Update content
@@ -94,22 +110,54 @@ class AnswerServiceImpl(
         // Remove photo deleted
         answer.photos.filter { !answerRequest.photos.contains(it.path) }
             .let {
-                answer.photos.removeAll(it)
+                answer.photos.removeAll(it.toSet())
                 photoRepository.deleteAll(it)
             }
 
         // Add photo, and update positions
-        answerRequest.photos.mapIndexed { index: Int, path: String ->
+        answerRequest.photos.forEachIndexed { index: Int, path: String ->
             answer.photos.find { it.path == path }
                 ?. let { // If photo exists, update its position
-                    answer.photos.remove(it)
-                    answer.photos.add(index, it)
+                    it.photosOrder = index
                 }
-                ?: PhotoEntity(path, answer = answer) // If photo doesn't exist, create new PhotoEntity and add to answer
-                    .also { answer.photos.add(index, it) }
+                ?: PhotoEntity(path, index, answer = answer) // If photo doesn't exist, create new PhotoEntity and add to answer
+                    .also { answer.photos.add(it) }
         }
 
         answerRepository.save(answer)
+    }
+
+    @Transactional
+    override fun toggleSelectAnswer(loginUser: UserEntity, answerId: Long, toSelect: Boolean) {
+        val answer = answerRepository.findByIdOrNull(answerId) // TODO: Optimize query (by fetch question-user)
+            ?: throw Jisik2n404("$answerId 답변이 존재하지 않습니다.")
+
+        if (loginUser.id != answer.question.user.id) {
+            throw Jisik2n403("질문한 사용자만 답변을 채택할 수 있습니다.")
+        }
+
+        if (answer.selected == toSelect) {
+            throw Jisik2n400("변경할 점이 없습니다.")
+        }
+
+        if (toSelect) { selectAnswer(answer) } else { unselectAnswer(answer) }
+    }
+
+    @Transactional
+    fun selectAnswer(answer: AnswerEntity) {
+        if (answer.question.close) {
+            throw Jisik2n403("마감된 질문에 답변을 채택할 수 없습니다.")
+        }
+
+        answer.selected = true
+        answer.selectedAt = LocalDateTime.now(ZoneId.of("Asia/Seoul"))
+        answer.question.close = true
+    }
+
+    @Transactional
+    fun unselectAnswer(answer: AnswerEntity) {
+        answer.selected = false
+        answer.question.close = false
     }
 
     @Transactional
@@ -117,9 +165,16 @@ class AnswerServiceImpl(
         val answer = answerRepository.findByIdOrNull(answerId)
 
         if (answer != null) {
-            if (answer.user.id != loginUser.id || answer.selected) {
-                TODO("Throw 403 Forbidden Exception")
+            if (answer.user.id != loginUser.id) {
+                throw Jisik2n403("자신의 게시물만 삭제할 수 있습니다.")
             }
+            if (answer.question.close) {
+                throw Jisik2n403("마감된 질문은 삭제될 수 없습니다.")
+            }
+
+            // Remove Interactions
+            userAnswerInteractionRepository.deleteByAnswer(answer)
+
             answerRepository.deleteById(answerId)
         }
     }
